+++ packages/node/src/subcommands/reindex.module.ts
@@ -5,7 +5,6 @@ import { Module } from '@nestjs/common';
 import { EventEmitter2, EventEmitterModule } from '@nestjs/event-emitter';
 import { SchedulerRegistry } from '@nestjs/schedule';
 import {
-  ApiService,
   DbModule,
   ForceCleanService,
   StoreService,
@@ -14,22 +13,21 @@ import {
   NodeConfig,
   ConnectionPoolService,
   ConnectionPoolStateManager,
+  UnfinalizedBlocksService,
+  DynamicDsService,
+  DsProcessorService,
 } from '@subql/node-core';
 import { Sequelize } from '@subql/x-sequelize';
+import { BlockchainService } from '../blockchain.service';
 import { ConfigureModule } from '../configure/configure.module';
-import { SubqueryProject } from '../configure/SubqueryProject';
-import { DsProcessorService } from '../indexer/ds-processor.service';
-import { DynamicDsService } from '../indexer/dynamic-ds.service';
-import { UnfinalizedBlocksService } from '../indexer/unfinalizedBlocks.service';
 import { StellarApiService } from '../stellar';
-import { StellarApiConnection } from '../stellar/api.connection';
 
 @Module({
   providers: [
     {
       provide: 'IStoreModelProvider',
       useFactory: storeModelFactory,
-      inject: [NodeConfig, EventEmitter2, SchedulerRegistry, Sequelize],
+      inject: [NodeConfig, EventEmitter2, Sequelize],
     },
     StoreService,
     ReindexService,
@@ -46,28 +44,18 @@ import { StellarApiConnection } from '../stellar/api.connection';
     ConnectionPoolService,
     ConnectionPoolStateManager,
     {
-      provide: ApiService,
-      useFactory: async (
-        project: SubqueryProject,
-        connectionPoolService: ConnectionPoolService<StellarApiConnection>,
-        eventEmitter: EventEmitter2,
-        nodeConfig: NodeConfig,
-      ) => {
-        const apiService = new StellarApiService(
-          project,
-          connectionPoolService,
-          eventEmitter,
-          nodeConfig,
-        );
-        await apiService.init();
-        return apiService;
-      },
+      provide: 'APIService',
+      useFactory: StellarApiService.create.bind(StellarApiService),
       inject: [
         'ISubqueryProject',
         ConnectionPoolService,
         EventEmitter2,
         NodeConfig,
       ],
+    },
+    {
+      provide: 'IBlockchainService',
+      useClass: BlockchainService,
     },
     SchedulerRegistry,
   ],