+++ packages/node/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Support http soroban endpoints (#109)
+- Update `@subql/node-core` to latest version and implement blockchain service as well as any relevant changes (#110)
+- Update nestjs dependencies (#110)
+- Update to Nodejs lts (#110)
 
 ## [4.5.0] - 2025-02-26
 ### Added